+++ packages/cad-simple-viewer/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @mlightcad/cad-simple-viewer
 
+## 1.0.2
+
+### Patch Changes
+
+- feat: refine MlCadViewer component by adding properties 'url' and 'wait'
+- Updated dependencies
+  - @mlightcad/svg-renderer@0.0.3
+  - @mlightcad/three-renderer@1.0.2
+
 ## 1.0.1
 
 ### Patch Changes