+++ packages/cad-viewer/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @mlightcad/cad-viewer
 
+## 1.0.2
+
+### Patch Changes
+
+- feat: refine MlCadViewer component by adding properties 'url' and 'wait'
+- Updated dependencies
+  - @mlightcad/cad-simple-viewer@1.0.2
+
 ## 1.0.1
 
 ### Patch Changes