+++ packages/svg-renderer/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @mlightcad/svg-renderer
 
+## 0.0.3
+
+### Patch Changes
+
+- feat: refine MlCadViewer component by adding properties 'url' and 'wait'
+
 ## 0.0.2
 
 ### Patch Changes